extract page states into seperate components
